pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "AIRenta"
version = "0.1.0"
description = "A telegram bot that automatically selects apartments according to your criteria and sends direct links to ads."
#readme = ""
requires-python = ">=3.11"
dependencies = [
    "celery>=5.3.0,<6.0.0",
    "aioredis>=2.0.0,<3.0.0",
    "aiogram>=3.0.0,<4.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "alembic>=1.12.0,<2.0.0",
    "fastapi[all]>=0.121.2",
    "uvicorn[standard]>=0.38.0",
    "aiohttp>=3.12.15",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.12.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0,<1.0.0",
    "pre-commit>=3.0.0,<4.0.0",
    "pytest>=7.0.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "httpx>=0.24.0,<1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]

# Настройки Ruff
[tool.ruff]
# Поддержка версии Python
target-version = "py311"

# Основные настройки
line-length = 120
indent-width = 4

# Настройки линтера
[tool.ruff.lint]
# Нестабильные и находящиеся в предварительной версии правила
preview = true
# Правила, которые должен применять линтер
select = [
    "UP",   # pyupgrade
    "I",    # isort
    "E",    # flake8-errors
    "W",    # flake8-warnings
    "F",    # flake8-pyflakes
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
    "T10",  # flake8-debugger
    "PT",   # flake8-pytest-style
    "RUF",  # Ruff
    "DTZ",  # flake8-datetimez
    "ERA",  # eradicate
    "FURB", # refurb
]
# Игнорирование конкретных предупреждений линтера
ignore = [
    "C408",    # Ненужный вызов dict, list и т.д.
    "RUF001",  # Игнорирование похожих символов с разными ASCII
    "RUF002",  # Игнорирование похожих символов с разными ASCII
    "RUF003",  # Игнорирование похожих символов с разными ASCII
    "S110",    # try-except-pass
]

# Ошибки, которые могут быть исправлены автоматически
fixable = [
    "I", # сортировка импортов
]

# Настройки форматтера
[tool.ruff.format]
quote-style = "single"
indent-style = "space"
